refactor: Migrate NoteEditorFragment to NoteEditorActivity
- Merges all logic, properties, view models, and helper methods from NoteEditorFragment directly into NoteEditorActivity.
- Deletes legacy XML layouts (note_editor.xml, note_editor_fragment.xml) and legacy NoteEditorFragment.kt file.
- Updates all navigation callers and dependencies throughout the codebase to reference NoteEditorActivity directly.
- Refactors instrumented tests (NoteEditorIntentTest and NoteEditorTabOrderTest) to use manual ActivityScenario launching within test methods, avoiding race conditions with CollectionManager initialization.
