+++ AnkiDroid/src/androidTest/java/com/ichi2/anki/testutil/NoteEditorActivity.kt
@@ -18,24 +18,21 @@ package com.ichi2.anki.testutil
 
 import androidx.test.core.app.ActivityScenario
 import com.ichi2.anki.NoteEditorActivity
-import com.ichi2.anki.NoteEditorFragment
-import com.ichi2.anki.R
 import java.util.concurrent.atomic.AtomicReference
 
 /**
- * Executes a block of code with the NoteEditor fragment on the activity's main thread.
- * @param block The block of code to execute with the NoteEditor fragment.
+ * Executes a block of code with the NoteEditor activity on the activity's main thread.
+ * @param block The block of code to execute with the NoteEditor activity.
  * @throws Throwable if any exception is thrown during the execution of the block.
  */
 @Throws(Throwable::class)
-fun ActivityScenario<NoteEditorActivity>.onNoteEditor(block: (NoteEditorFragment) -> Unit) {
+fun ActivityScenario<NoteEditorActivity>.onNoteEditor(block: (NoteEditorActivity) -> Unit) {
     val wrapped = AtomicReference<Throwable?>(null)
     this.onActivity { activity: NoteEditorActivity ->
         try {
-            val editor: NoteEditorFragment = activity.getNoteEditorFragment()
             activity.runOnUiThread {
                 try {
-                    block(editor)
+                    block(activity)
                 } catch (t: Throwable) {
                     wrapped.set(t)
                 }
@@ -46,6 +43,3 @@ fun ActivityScenario<NoteEditorActivity>.onNoteEditor(block: (NoteEditorFragment
     }
     wrapped.get()?.let { throw it }
 }
-
-fun NoteEditorActivity.getNoteEditorFragment(): NoteEditorFragment =
-    supportFragmentManager.findFragmentById(R.id.note_editor_fragment_frame) as NoteEditorFragment
