+++ AnkiDroid/src/main/java/com/ichi2/anki/DeckPicker.kt
@@ -264,9 +264,9 @@ open class DeckPicker : AnkiActivity(), SyncErrorDialogListener, ImportDialogLis
             }
             val data = result.data
             if (data != null && (data.getBooleanExtra(
-                    NoteEditorFragment.RELOAD_REQUIRED_EXTRA_KEY,
+                    NoteEditorActivity.RELOAD_REQUIRED_EXTRA_KEY,
                     false,
-                ) || data.getBooleanExtra(NoteEditorFragment.NOTE_CHANGED_EXTRA_KEY, false))
+                ) || data.getBooleanExtra(NoteEditorActivity.NOTE_CHANGED_EXTRA_KEY, false))
             ) {
                 cardBrowserViewModel.search(cardBrowserViewModel.searchQuery.value)
             }
