+++ AnkiDroid/src/test/java/com/ichi2/anki/browser/CardBrowserViewModelTest.kt
@@ -28,7 +28,6 @@ import com.ichi2.anki.AnkiDroidApp
 import com.ichi2.anki.CollectionManager
 import com.ichi2.anki.Flag
 import com.ichi2.anki.NoteEditorActivity
-import com.ichi2.anki.NoteEditorFragment
 import com.ichi2.anki.R
 import com.ichi2.anki.browser.CardBrowserColumn.ANSWER
 import com.ichi2.anki.browser.CardBrowserColumn.CARD
@@ -180,7 +179,7 @@ class CardBrowserViewModelTest : JvmTest() {
 
         val addIntent = com.ichi2.anki.noteeditor.NoteEditorLauncher.AddNoteFromCardBrowser(this).toIntent(mockIt())
         val bundle = addIntent.getBundleExtra(NoteEditorActivity.FRAGMENT_ARGS_EXTRA)
-        IntentAssert.doesNotHaveExtra(bundle, NoteEditorFragment.EXTRA_DID)
+        IntentAssert.doesNotHaveExtra(bundle, NoteEditorActivity.EXTRA_DID)
     }
 
     @Test
